Fix undefined behavior in the SID filter, and measure what we can see
UBSan found a left shift of a negative value in filter_clock. On this
hardware it computed the intended result, but C11 leaves it undefined,
so the code was correct in practice and not correct in principle. The
scale is now done in an unsigned type and converted back, which the
standard defines and which produces the same bit pattern.

Proven behavior-preserving: a deterministic filter run driving the exact
UB path (routed voices, cutoff sweep, resonance steps, all modes, 30000
cycles) hashes to fa8275ba906e3588 both before and after the fix. Not one
sample changed. That run is now a durable test, mutation-checked, so the
filter's arithmetic cannot silently alter the sound again.

Scanned the rest of the codebase for the same class: line 398 was the
only signed left shift that can take a negative operand. Every other
shift is on an unsigned type or a value that cannot be negative. UBSan
agrees across every executed path.

Added make test-asan (address and undefined sanitizers, gating) and make
coverage (per-file line coverage). Coverage: vic 90%, mem 92%, cia 89%,
sid 84%, bus 84%, drive 85%, cpu6502 60% without Lorenz, host 0% because
the SDL layer never runs headless.

This is not a claim of memory safety. The tools only see what the tests
execute. What they found on those paths was one defect, now fixed.

+++ src/sid.c
@@ -395,7 +395,9 @@ static void filter_clock(void) {
             direct += s;  // voice 3 can be cut from the direct path (3OFF)
         }
     }
-    int32_t in = routed << FILT_STATE_SHIFT;
+    // Scale in unsigned: a left shift of a negative signed value is undefined in
+    // C; the conversion back is the two's-complement bit pattern we want.
+    int32_t in = (int32_t)((uint32_t)routed << FILT_STATE_SHIFT);
     int32_t lp = filt.lp;
     int32_t bp = filt.bp;
     lp += (int32_t)(((int64_t)filt.coef_f * bp) >> FILT_COEF_SHIFT);
